crawler/pyproject.toml
[tool.poetry]
name = "crawler"
version = "0.1.0"
description = ""
authors = ["Gongmeda <gongmeda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
openai = "^0.27.6"
feedparser = "^6.0.10"
beautifulsoup4 = "^4.12.2"
apscheduler = "^3.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
